Fix buffer-package for non-standard package value
If (buffer-value buffer "package") returned a value that was not
a string designator, and error was signaled and Lem died.
E.g., the file "iterate.lisp" from project :iterate contains the
following property list:

;;;-*- syntax:COMMON-LISP; Package: (ITERATE :use "COMMON-LISP" :colon-mode :external) -*-

Therefore, openning that file and calling 'lisp-mode' killed Lem.
Now, wrap call with ignore-errors and use default if NIL is returned.