Fixed an issue when connection to swank didn't initialized propertly
The root of the problem was that on OSX swank:*communication-style* is equal to :fd-handler
for some reason. In this case swank:create-server does not start threads which
accept connections on TCP port and process incoming messages.

To make it work propertly, we need to specify that communication style is :spawn.

This commit closes issue https://github.com/cxxxr/lem/issues/411.