Allow a multi-key sequence to override a single-key definition
Currently, if a key is already bound, attempting to override that
binding produces an error, specifically, if the following config code
is used:

```lisp
(define-key lem-lisp-mode:*lisp-mode-keymap* "," 'my-function)
(define-key lem-lisp-mode:*lisp-mode-keymap* ", ," 'my-other-function)
```

```
The value
  MY_FUNCTION
is not of type
  HASH-TABLE
when binding HASH-TABLE
```

Which is not particularly helpful.

This commit adjusts the behavior of `define-key` in the above
situation: change the definition of the key to the given definition,
regardless of whether it's already been defined.