Remove lem-legit:*vcs* by making users take it as a function parameter
Yes, Emacs makes extensive use of dynamic bindings/scope: that's not a
reason we should do the same. Common Lisp supports lexical scoping,
unlike Elisp when it started out, and in my opinion we should really,
*really*, try to avoid dynamic binds where a function parameter would
do.

Yes, this adds the need to plumb an extra parameter around. That's well
worth the cost: we can always package/split function parameters as
needed and as they make sense, whereas dynamic bindings may prove to be
a *huge* pain point in the future if Lem starts making use of paralellism.

The Emacs documentation itself puts it succinctly [1]

> Lexical binding opens up many more opportunities for optimization,
> so programs using it are likely to run faster in future Emacs versions.
> Lexical binding is also more compatible with concurrency,

Now, focusing on Legit itself: much of the in lem/porcelain uses the
pattern

```
(defun some-legit-function (vcs)
  (case vcs
    (:git (git-fn))
    (:hg (hg-fn))
    (:t (fallthrough))))
```

There's no reason to have all of this dispatch logic done manually: this
is a textbook case for using generic functions. Now, the thing is that
generic functions need something to dispatch *on*. And it's pretty clear
that the VCS parameter (representing, really, the repository [2]) is the
perfect thing for that.

By switching all the `legit-function`s to be generic functions, the
implementation for each VCS can also define their `vcs` objects to
contain whatever data they might find useful: whether that's some sort
of legit-side cache for commits, FFI handles (will be useful for libgit2
support for instance), etc.

[1] https://www.gnu.org/software/emacs/manual/html_node/elisp/Lexical-Binding.html
[2] Might be worth a rename, but I don't want to clutter this particular
change.