Add SLIME's C-c ~ command (sync package) to lisp-mode
In Emacs, pressing C-c ~ in a Lisp buffer with SLIME will cause
the REPL's package to be switched to the buffer's active package.
This change implements that functionality.

Note that it doesn't have the functionality of switching Lisp's
current directory. This is partially because I found that
behaviour janky in Lem, and partially because I don't like it
doing that in Emacs either.