at present, Lem's `kill-whole-line` kills the line and then does `delete-previous-char`, which seems wrong.  this patch makes `kill-whole-line`, one, kill the entire line, including the new line at the end of a line if indeed there is one, and, two, allow it to be prefixed.
|;; one one one
;; two two two
;; three three three
;; four four four
;; five five five

(| is point) if `5 kill-whole-line`, all lines are deleted.

;; one one one
;; two two two
;; three three |three
;; four four four
;; five five five

(| is point) if `0 kill-whole-line`, nothing happens.

;; one one one
;; two two two
;; three three |three
;; four four four
;; five five five

(| is point) if `-2 kill-whole-line`, ";; one one one" and ";; two two two" are deleted, and the point stays right before the last "three".  if you want to delete the previous "three"s, then you can use the `kill` command.

i'm curious if others agree with me