update point after backspacing and yanking while isearch is active
|
;; one test is what you need

(| is the point) with this patch, if you search for "ee" and then backspace, the point will move to the end of the word "one".   in Lem currently, if you search for "ee" and then backspace, the point will stay after "ee", when I feel it should be at the end of "one" since the search string has changed.  this patch also works when reverse searching

|
;; one test is what you need

(| is the point)  with this patch, if  "one test" is the current kill string and you yank, the point will move to after "one test".  in Lem currently, the point will not move after yanking, which i don't think is natural