change `move-over-)` to `move-over-)-or-wrap` so that if the mark is active, the region will be wrapped with parens and the point will be placed after the close paren, helpful if you don't need to add anything in the parens. could also be helpful if you use pareto mode. and the name is changed so that functionality is not hidden from the user
an odd command, but one that seems like a logical counterpart to insert-()-or-wrap

the difference between this patch and `insert-()-or-wrap` is that this one puts the point at the end, after the close paren.

(when (point> end |current-point)

(| is the point.)  if we put the region around "current-point" and then execute this command, we will get "(when (point> end (current-point))" and then point will be to the right of the close paren.

(when (point> end (current-point)|)

(| is the point) and then we can execute this command (the region will not be active), and we will skip over the close paren and go to the next line.

(when (point> end (current-point))
  |

(| is the point.)  seems that it could be useful.