add two editor variables to line-numbers
this patch adds two editor variables to line-numbers, `line-format` and `current-line-display-function`.  this allows the user to change the width/format of the line numbers and to set the current-line display in relative numbers.

for example :
```
(setf (variable-value 'line-number-format :global) "~2D ")
(setf (variable-value 'current-line-display-function :global) (lambda () 0))
```

now the line-numbers will be narrower and the current-line will always be 0.