only change recent frame list when switch target is different
only change recent frame list when switch target is different.  i find this to be very helpful because sometimes i choose the same frame either through a command or by accident, and then `frame-multiplexer-recent` becomes useless.  with this change, `frame-multiplexer-recent` will continue to be useful even after a mitake or a redundant switch

for example :
i run a command that always opens a file in frame 1.  so i run the command a few times (maybe while changing buffers on the same frame), and now when i do `frame-multiplexer-recent`, it is useless.  with this patch, `frame-multiplexer-recent` will continue to be useful in such situations