directory-mode: follow buffer name and directory when moving one file (r) (#1945)
* directory-mode: rename file and buffer name (r)

for https://github.com/lem-project/lem/issues/993

* directory-mode: add "New name" prompt to inline rename prompt (r)

for https://github.com/lem-project/lem/issues/993

* directory-mode: don't move multiple files to 1 file ("R")

Previously: select multiple files, press "R", choose a
destination *filename* (and not a directory): we loose a file.

Now, if we select many files, we check that the destination is a directory.

for https://github.com/lem-project/lem/issues/993