Adds a new bookmark package.
Bookmarks are stored file paths, possibly with a location, which enables
the user to quickly visit files which are often needed.  Bookmark
information are kept in-memory, the user has to use the commands
BOOKMARK-LOAD and BOOKMARK-SAVE manually to persist bookmark
information.

The command BOOKMARK-SET can be used to create a new bookmark for the
current buffer with the current cursor position.  To go to the location
of a previously created bookmark, the command BOOKMARK-JUMP can be used.
More commands are available to manage bookmarks, see the exported
symbols from the package.